pyproject.toml
[project]
name = "crewai-data-analyst"
version = "0.1.0"
description = "CrewAI agent for data analysis with vnstock integration"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.30.0",
    "crewai-tools>=0.4.0",
    "streamlit>=1.47.1",
    "vnstock>=3.2.6",
    "pandas>=2.0.0",
    "pandasai>=2.0.0",
    "openai>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
the-crew-2="crewai_data_analyst.main:kickoff"
run_crew = "crewai_data_analyst.main:kickoff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crewai_data_analyst"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[tool.crewai]
type = "crew"